+++ main.js
@@ -1109,6 +1109,7 @@ app.whenReady().then(async () => {
     logger = new Logger();
     dnrBridge.setLogger(logger);
     logger.info('main', '=== BunPinokWeb запущен v' + app.getVersion() + ' ===');
+    logger.info('main', '=== BunPinokWeb запущен v' + app.getVersion() + ' ===');
     permissionDialogManager = new PermissionDialogManager();
     passwordStore = new PasswordStore();
     authDialogManager = new AuthDialogManager();
@@ -1183,6 +1184,22 @@ app.whenReady().then(async () => {
         }
     });
     tabManager.setChromeExtensions(chromeExtensions);
+    dnrBridge.setContext({
+        bookmarkStore: bookmarkStore,
+        historyStore: historyStore,
+        tabManager: tabManager,
+        broadcastBookmarks: () => {
+            if (mainWindow && !mainWindow.isDestroyed()) {
+                mainWindow.webContents.send('bookmarks:updated', bookmarkStore.getAll());
+            }
+        },
+        updateBounds: () => {
+            updateChromeViewBounds();
+        },
+        applyProxy: () => {
+            applyProxySettings();
+        }
+    });
     ElectronChromeExtensions.handleCRXProtocol(session.defaultSession);
     console.log('[ElectronChromeExtensions] инициализированы');